+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* `BooleanFilterOptionList.maxNumToots` is now a variable not a function
 
 ### v0.64.2
 * (Demo App) Bump `webpack-dev-server` to version 5.22 via `npm audit force`